Sources were synced with upstream: Psi was updated. Iris was updated. Plugins were updated.

+++ iris/src/xmpp/xmpp-im/filetransfer.cpp
@@ -14,7 +14,7 @@
  *
  * You should have received a copy of the GNU Lesser General Public
  * License along with this library; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
  *
  */
 
@@ -642,7 +642,7 @@ bool JT_FT::take(const QDomElement &x)
 		if(!file.isNull()) {
 			QDomElement range = file.elementsByTagName("range").item(0).toElement();
 			if(!range.isNull()) {
-				int x;
+				qlonglong x;
 				bool ok;
 				if(range.hasAttribute("offset")) {
 					x = range.attribute("offset").toLongLong(&ok);
